Move use-file to hooks folder
